requirements.txt
# Heterogeneous TP Configuration Testing Framework Dependencies

# Core ML/DL frameworks
# vLLM 0.9.0 requires PyTorch 2.4+ and transformers 4.45.x
vllm==0.9.0
torch>=2.4.0,<2.6.0
transformers==4.51.1

# Dataset handling
datasets>=2.18.0,<3.0.0

# GPU monitoring
pynvml>=11.5.0
nvidia-ml-py>=12.555.0

# Data processing and analysis
pandas>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0

# SciPy (pin version to avoid NumPy conflicts)
scipy>=1.11.0,<1.14.0

# Visualization
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0

# Configuration
pyyaml>=6.0.1

# Async HTTP client
aiohttp>=3.9.0,<4.0.0
httpx>=0.27.0,<1.0.0

# Progress bars
tqdm>=4.66.0

# Utilities
python-dateutil>=2.8.2
psutil>=5.9.8

# Logging
colorlog>=6.8.0

# Testing (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
